+++ src/main/java/com/repository/Purchase.java
@@ -32,7 +32,7 @@ public class Purchase implements Serializable{
     @Temporal(TemporalType.TIMESTAMP)
     private Date purchaseDate;
 
-    private Float finalPrice;
+    private Double finalPrice;
 
     public Purchase() {
         games = new ArrayList<>();
